requirements.txt
attrs==19.3.0
Automat==0.8.0
backcall==0.1.0
cffi==1.13.1
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
ipdb==0.12.2
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
lxml==4.4.1
parsel==1.5.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
python-slugify==4.0.0
queuelib==1.5.0
Scrapy==1.7.4
service-identity==18.1.0
six==1.12.0
text-unidecode==1.3
traitlets==4.3.3
Twisted==19.7.0
w3lib==1.21.0
wcwidth==0.1.7
zope.interface==4.6.0
